+++ internal/resources/common/constructors/graph_beta/device_and_app_management/mobile_application_assignments.go
