+++ internal/resources/common/crud/graph_beta/device_and_app_management/mobile_application_helpers.go
@@ -1,4 +1,4 @@
-package graphBetaMacOSPKGApp
+package helpers
 
 import (
 	"context"
@@ -14,8 +14,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	msgraphbetasdk "github.com/microsoftgraph/msgraph-beta-sdk-go"
-	graphmodels "github.com/microsoftgraph/msgraph-beta-sdk-go/models"
 )
 
 // TempFileInfo holds information about a temporary file that needs to be cleaned up
@@ -26,9 +24,9 @@ type TempFileInfo struct {
 	ShouldCleanup bool
 }
 
-// setInstallerSourcePath inspects the resource model's app_metadata and returns the actual installer file path
+// SetInstallerSourcePath inspects the resource model's app_installer and returns the actual installer file path
 // along with information about whether it needs cleanup.
-func setInstallerSourcePath(ctx context.Context, metadataObj types.Object) (string, TempFileInfo, error) {
+func SetInstallerSourcePath(ctx context.Context, metadataObj types.Object) (string, TempFileInfo, error) {
 	var fileInfo TempFileInfo
 	var metadata sharedmodels.MobileAppMetaDataResourceModel
 
@@ -36,7 +34,7 @@ func setInstallerSourcePath(ctx context.Context, metadataObj types.Object) (stri
 	if !metadataObj.IsNull() {
 		diags := metadataObj.As(ctx, &metadata, basetypes.ObjectAsOptions{})
 		if diags.HasError() {
-			tflog.Warn(ctx, "Failed to parse app_metadata, proceeding with empty struct", map[string]interface{}{
+			tflog.Warn(ctx, "Failed to parse app_installer, proceeding with empty struct", map[string]interface{}{
 				"errors": diags.Errors(),
 			})
 		}
@@ -78,8 +76,8 @@ func setInstallerSourcePath(ctx context.Context, metadataObj types.Object) (stri
 	return "", fileInfo, fmt.Errorf("installer file not provided; please supply either a local file path or a URL")
 }
 
-// cleanupTempFile removes a temporary file if it exists and should be cleaned up
-func cleanupTempFile(ctx context.Context, fileInfo TempFileInfo) {
+// CleanupTempFile removes a temporary file if it exists and should be cleaned up
+func CleanupTempFile(ctx context.Context, fileInfo TempFileInfo) {
 	if !fileInfo.ShouldCleanup || fileInfo.FilePath == "" {
 		return
 	}
@@ -141,103 +139,4 @@ func GetAppMetadata(
 		InstallerMD5Checksum:    types.StringValue(md5Checksum),
 		InstallerSHA256Checksum: types.StringValue(sha256Checksum),
 	}, nil
-}
-
-// evaluateIfContentVersionUpdateRequired determines if a new content version needs to be created and uploaded
-// based on source path changes and metadata differences
-func evaluateIfContentVersionUpdateRequired(ctx context.Context, object *MacOSPKGAppResourceModel, state *MacOSPKGAppResourceModel, client *msgraphbetasdk.GraphServiceClient) (bool, string, error) {
-
-	pathChanged := false
-	metadataChanged := false
-	existingContentVersion := ""
-
-	// Check installer source path changes
-	var objectMetadata, stateMetadata sharedmodels.MobileAppMetaDataResourceModel
-
-	if !object.AppMetadata.IsNull() {
-		diags := object.AppMetadata.As(ctx, &objectMetadata, basetypes.ObjectAsOptions{})
-		if diags.HasError() {
-			return false, "", fmt.Errorf("failed to parse object metadata: %v", diags)
-		}
-	}
-
-	if !state.AppMetadata.IsNull() {
-		diags := state.AppMetadata.As(ctx, &stateMetadata, basetypes.ObjectAsOptions{})
-		if diags.HasError() {
-			return false, "", fmt.Errorf("failed to parse state metadata: %v", diags)
-		}
-	}
-
-	// Path changes
-	if !objectMetadata.InstallerFilePathSource.Equal(stateMetadata.InstallerFilePathSource) {
-		tflog.Debug(ctx, "Installer file path has changed", map[string]interface{}{
-			"old_path": stateMetadata.InstallerFilePathSource.ValueString(),
-			"new_path": objectMetadata.InstallerFilePathSource.ValueString(),
-		})
-		pathChanged = true
-	}
-
-	if !objectMetadata.InstallerURLSource.Equal(stateMetadata.InstallerURLSource) {
-		tflog.Debug(ctx, "Installer URL has changed", map[string]interface{}{
-			"old_url": stateMetadata.InstallerURLSource.ValueString(),
-			"new_url": objectMetadata.InstallerURLSource.ValueString(),
-		})
-		pathChanged = true
-	}
-
-	// If no installer source is provided at all → skip content evaluation
-	if (objectMetadata.InstallerFilePathSource.IsNull() || objectMetadata.InstallerFilePathSource.ValueString() == "") &&
-		(objectMetadata.InstallerURLSource.IsNull() || objectMetadata.InstallerURLSource.ValueString() == "") {
-		tflog.Debug(ctx, "No installer source provided, skipping content evaluation")
-		return false, "", nil
-	}
-
-	// Check metadata changes
-	if !object.AppMetadata.IsNull() && !state.AppMetadata.IsNull() {
-		if objectMetadata.InstallerSizeInBytes.ValueInt64() != stateMetadata.InstallerSizeInBytes.ValueInt64() {
-			tflog.Debug(ctx, "Installer size has changed", map[string]interface{}{
-				"previous_size": stateMetadata.InstallerSizeInBytes.ValueInt64(),
-				"current_size":  objectMetadata.InstallerSizeInBytes.ValueInt64(),
-			})
-			metadataChanged = true
-		}
-
-		if objectMetadata.InstallerSHA256Checksum.ValueString() != stateMetadata.InstallerSHA256Checksum.ValueString() {
-			tflog.Debug(ctx, "Installer SHA256 checksum has changed", map[string]interface{}{
-				"previous_checksum": stateMetadata.InstallerSHA256Checksum.ValueString(),
-				"current_checksum":  objectMetadata.InstallerSHA256Checksum.ValueString(),
-			})
-			metadataChanged = true
-		}
-	} else {
-		tflog.Debug(ctx, "Metadata is missing from one of the objects, assuming content needs update")
-		metadataChanged = true
-	}
-
-	// Retrieve existing committed content version (from API)
-	resource, err := client.
-		DeviceAppManagement().
-		MobileApps().
-		ByMobileAppId(object.ID.ValueString()).
-		Get(ctx, nil)
-
-	if err == nil {
-		macOSPkgApp, ok := resource.(graphmodels.MacOSPkgAppable)
-		if ok && macOSPkgApp.GetCommittedContentVersion() != nil &&
-			*macOSPkgApp.GetCommittedContentVersion() != "" {
-			existingContentVersion = *macOSPkgApp.GetCommittedContentVersion()
-			tflog.Debug(ctx, fmt.Sprintf("Found existing committed content version from API: %s", existingContentVersion))
-		}
-	}
-
-	updateNeeded := pathChanged || metadataChanged
-
-	tflog.Debug(ctx, "Content version update evaluation result", map[string]interface{}{
-		"update_needed":            updateNeeded,
-		"path_changed":             pathChanged,
-		"metadata_changed":         metadataChanged,
-		"existing_content_version": existingContentVersion,
-	})
-
-	return updateNeeded, existingContentVersion, nil
 }
