+++ internal/resources/common/state/graph_beta/device_and_app_management/mobile_application_assignments.go
