+++ internal/resources/common/state/graph_beta/device_and_app_management/mobile_application_metadata.go
@@ -1,4 +1,4 @@
-package graphBetaMacOSPKGApp
+package sharedStater
 
 import (
 	"context"
