+++ internal/resources/device_and_app_management/graph_beta/macos_pkg_app/crud.go
@@ -3,16 +3,19 @@ package graphBetaMacOSPKGApp
 import (
 	"context"
 	"fmt"
+	"path/filepath"
 	"time"
 
 	construct "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/constructors/graph_beta/device_and_app_management"
 	"github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/crud"
+	helpers "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/crud/graph_beta/device_and_app_management"
 	"github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/errors"
 	sharedmodels "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/shared_models/graph_beta/device_and_app_management"
 	sharedstater "github.com/deploymenttheory/terraform-provider-microsoft365/internal/resources/common/state/graph_beta/device_and_app_management"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/microsoftgraph/msgraph-beta-sdk-go/deviceappmanagement"
@@ -62,7 +65,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 	retryTimeout := time.Until(deadline) - time.Second
 
 	// Step 1: Determine installer source path (local or download via URL)
-	installerSourcePath, tempFileInfo, err := setInstallerSourcePath(ctx, object.AppMetadata)
+	installerSourcePath, tempFileInfo, err := helpers.SetInstallerSourcePath(ctx, object.AppInstaller)
 	if err != nil {
 		resp.Diagnostics.AddError("Error determining installer file path", err.Error())
 		return
@@ -70,7 +73,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 
 	// Ensure cleanup of temporary file occurs post state read
 	if tempFileInfo.ShouldCleanup {
-		defer cleanupTempFile(ctx, tempFileInfo)
+		defer helpers.CleanupTempFile(ctx, tempFileInfo)
 	}
 
 	// Step 3: Construct the base resource from the Terraform model
@@ -104,9 +107,9 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 			return
 		}
 
-		err = assignCategoriesToMobileApplication(ctx, r.client, object.ID.ValueString(), categoryValues, r.ReadPermissions)
+		err = construct.AssignMobileAppCategories(ctx, r.client, object.ID.ValueString(), categoryValues, r.ReadPermissions)
 		if err != nil {
-			errors.HandleGraphError(ctx, err, resp, "Associate Categories", r.WritePermissions)
+			errors.HandleGraphError(ctx, err, resp, "Create", r.WritePermissions)
 			return
 		}
 	}
@@ -134,7 +137,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 
 		// Step 7: Encrypt mobile app file and output file and encryption metadata
 		tflog.Debug(ctx, "Encrypting installer file and constructing file metadata")
-		contentFile, encryptionInfo, err := encryptMobileAppAndConstructFileContentMetadata(ctx, installerSourcePath)
+		contentFile, encryptionInfo, err := construct.EncryptMobileAppAndConstructFileContentMetadata(ctx, installerSourcePath)
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Error constructing and encrypting Intune Mobile App content file",
@@ -215,7 +218,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 		// IMPORTANT: Upload the encrypted file (.bin) not the original source file.
 		encryptedFilePath := installerSourcePath + ".bin"
 		tflog.Debug(ctx, fmt.Sprintf("Uploading encrypted file: %s", encryptedFilePath))
-		err = uploadToAzureStorage(ctx, *fileStatus.GetAzureStorageUri(), encryptedFilePath)
+		err = construct.UploadToAzureStorage(ctx, *fileStatus.GetAzureStorageUri(), encryptedFilePath)
 		if err != nil {
 			tflog.Debug(ctx, fmt.Sprintf("Failed to upload to Azure Storage: %v", err))
 			errors.HandleGraphError(ctx, err, resp, "Create", r.WritePermissions)
@@ -225,7 +228,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 		// Step 11: Commit the file with encryption metadata
 		tflog.Debug(ctx, "Committing file with encryption metadata")
 		err = retry.RetryContext(ctx, time.Until(deadline), func() *retry.RetryError {
-			commitBody, err := CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
+			commitBody, err := construct.CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
 			if err != nil {
 				return retry.NonRetryableError(fmt.Errorf("failed to construct commit request: %v", err))
 			}
@@ -271,7 +274,7 @@ func (r *MacOSPKGAppResource) Create(ctx context.Context, req resource.CreateReq
 			}
 			if state == graphmodels.COMMITFILEFAILED_MOBILEAPPCONTENTFILEUPLOADSTATE {
 				tflog.Debug(ctx, "File commit failed; retrying commit request")
-				commitBody, err := CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
+				commitBody, err := construct.CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
 				if err != nil {
 					tflog.Debug(ctx, fmt.Sprintf("Error constructing commit request during retry: %v", err))
 					continue
@@ -426,27 +429,12 @@ func (r *MacOSPKGAppResource) Read(ctx context.Context, req resource.ReadRequest
 
 	MapRemoteResourceStateToTerraform(ctx, &object, macOSPkgApp)
 
-	// --- Added Debug Logging ---
+	// 2. get committed app content file versions
 	committedVersionIdPtr := macOSPkgApp.GetCommittedContentVersion()
-	if committedVersionIdPtr == nil {
-		tflog.Debug(ctx, fmt.Sprintf("App ID %s: GetCommittedContentVersion() returned nil", object.ID.ValueString()))
-	} else {
-		tflog.Debug(ctx, fmt.Sprintf("App ID %s: GetCommittedContentVersion() returned value: '%s'", object.ID.ValueString(), *committedVersionIdPtr))
-	}
-	// --- End Debug Logging ---
 
-	// Check if the pointer is not nil AND the dereferenced string is not empty (Corrected Logic)
 	if committedVersionIdPtr != nil && *committedVersionIdPtr != "" {
-		committedVersionId := *committedVersionIdPtr // Now it's safe to dereference
-		tflog.Debug(ctx, fmt.Sprintf("App ID %s: Found committed content version ID: %s. Proceeding to fetch files.", object.ID.ValueString(), committedVersionId))
-	}
+		committedVersionId := *committedVersionIdPtr
 
-	// 2. Get only the committed content version and its files
-	if macOSPkgApp.GetCommittedContentVersion() == nil || *macOSPkgApp.GetCommittedContentVersion() == "" {
-		committedVersionId := *macOSPkgApp.GetCommittedContentVersion()
-		tflog.Debug(ctx, fmt.Sprintf("Found committed content version ID: %s", committedVersionId))
-
-		// Get files for the committed version
 		respFiles, err := r.client.DeviceAppManagement().
 			MobileApps().
 			ByMobileAppId(object.ID.ValueString()).
@@ -461,8 +449,23 @@ func (r *MacOSPKGAppResource) Read(ctx context.Context, req resource.ReadRequest
 			return
 		}
 
-		// Map directly to Terraform state with the consolidated function
-		object.ContentVersion = MapCommittedContentVersionStateToTerraform(ctx, committedVersionId, respFiles, err)
+		var installerFileName string
+		if macOSPkgApp.GetFileName() != nil {
+			installerFileName = *macOSPkgApp.GetFileName()
+		} else {
+			var metadataObj sharedmodels.MobileAppMetaDataResourceModel
+			if !object.AppInstaller.IsNull() {
+				diags := object.AppInstaller.As(ctx, &metadataObj, basetypes.ObjectAsOptions{})
+				if !diags.HasError() && !metadataObj.InstallerFilePathSource.IsNull() {
+					filePath := metadataObj.InstallerFilePathSource.ValueString()
+					if filePath != "" {
+						installerFileName = filepath.Base(filePath)
+					}
+				}
+			}
+		}
+
+		object.ContentVersion = sharedstater.MapCommittedContentVersionStateToTerraform(ctx, committedVersionId, respFiles, err, installerFileName)
 	}
 
 	// 3. app assignments
@@ -481,31 +484,31 @@ func (r *MacOSPKGAppResource) Read(ctx context.Context, req resource.ReadRequest
 	object.Assignments = sharedstater.StateMobileAppAssignment(ctx, nil, respAssignments)
 
 	// 4. Get app metadata by processing app installer file
-	installerPath, tempInfo, err := setInstallerSourcePath(ctx, object.AppMetadata)
+	installerPath, tempInfo, err := helpers.SetInstallerSourcePath(ctx, object.AppInstaller)
 	if err != nil {
 		resp.Diagnostics.AddError("Error determining installer path", err.Error())
 		return
 	}
 	if tempInfo.ShouldCleanup {
-		defer cleanupTempFile(ctx, tempInfo)
+		defer helpers.CleanupTempFile(ctx, tempInfo)
 	}
 
 	var existingMetadata sharedmodels.MobileAppMetaDataResourceModel
 	if !req.State.Raw.IsNull() {
-		diags := req.State.GetAttribute(ctx, path.Root("app_metadata"), &existingMetadata)
+		diags := req.State.GetAttribute(ctx, path.Root("app_installer"), &existingMetadata)
 		if diags.HasError() {
 			resp.Diagnostics.Append(diags...)
 			return
 		}
 	}
 
-	metadata, err := GetAppMetadata(ctx, installerPath, &existingMetadata)
+	metadata, err := helpers.GetAppMetadata(ctx, installerPath, &existingMetadata)
 	if err != nil {
-		resp.Diagnostics.AddError("Error capturing installer metadata", err.Error())
+		resp.Diagnostics.AddError("Error capturing app installer metadata", err.Error())
 		return
 	}
 
-	object.AppMetadata = MapAppMetadataStateToTerraform(ctx, metadata)
+	object.AppInstaller = sharedstater.MapAppMetadataStateToTerraform(ctx, metadata)
 
 	// 6. set final state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &object)...)
@@ -520,7 +523,7 @@ func (r *MacOSPKGAppResource) Read(ctx context.Context, req resource.ReadRequest
 func (r *MacOSPKGAppResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 	var object, state MacOSPKGAppResourceModel
 	var installerSourcePath string
-	var tempFileInfo TempFileInfo
+	var tempFileInfo helpers.TempFileInfo
 	var err error
 
 	tflog.Debug(ctx, fmt.Sprintf("Starting Update of resource: %s_%s", r.ProviderTypeName, r.TypeName))
@@ -539,13 +542,11 @@ func (r *MacOSPKGAppResource) Update(ctx context.Context, req resource.UpdateReq
 
 	deadline, _ := ctx.Deadline()
 	retryTimeout := time.Until(deadline) - time.Second
-
 	// Ensure cleanup of temporary file when we're done
 	if tempFileInfo.ShouldCleanup {
-		defer cleanupTempFile(ctx, tempFileInfo)
+		defer helpers.CleanupTempFile(ctx, tempFileInfo)
 	}
-
-	installerSourcePath, tempFileInfo, err = setInstallerSourcePath(ctx, object.AppMetadata)
+	installerSourcePath, tempFileInfo, err = helpers.SetInstallerSourcePath(ctx, object.AppInstaller)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error determining installer file path",
@@ -586,7 +587,7 @@ func (r *MacOSPKGAppResource) Update(ctx context.Context, req resource.UpdateReq
 			return
 		}
 
-		err = assignCategoriesToMobileApplication(ctx, r.client, object.ID.ValueString(), categoryValues, r.ReadPermissions)
+		err = construct.AssignMobileAppCategories(ctx, r.client, object.ID.ValueString(), categoryValues, r.ReadPermissions)
 		if err != nil {
 			errors.HandleGraphError(ctx, err, resp, "Update Categories", r.WritePermissions)
 			return
@@ -623,243 +624,27 @@ func (r *MacOSPKGAppResource) Update(ctx context.Context, req resource.UpdateReq
 			return
 		}
 	}
-
-	// Step 4:  Updated application content version and files
-	if installerSourcePath != "" {
-
-		// Evaluate if content update is needed
-		contentUpdateNeeded, existingContentVersion, err := evaluateIfContentVersionUpdateRequired(ctx, &object, &state, r.client)
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error evaluating if content update is needed",
-				err.Error(),
-			)
-			return
-		}
-
-		// --- Begin File Processing Logic ---
-		// Process content upload only if content has changed or we need a new version
-		if contentUpdateNeeded {
-			tflog.Debug(ctx, "Content has changed — initializing new content version for upload")
-
-			// Step 4b: Initialize new content version for file upload
-			content := graphmodels.NewMobileAppContent()
-			contentBuilder := r.client.
-				DeviceAppManagement().
-				MobileApps().
-				ByMobileAppId(object.ID.ValueString()).
-				GraphMacOSPkgApp().
-				ContentVersions()
-
-			contentVersion, err := contentBuilder.Post(ctx, content, nil)
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-			tflog.Debug(ctx, fmt.Sprintf("New content version created with ID: %s", *contentVersion.GetId()))
-
-			// Step 4c: Encrypt installer file and construct file metadata
-			tflog.Debug(ctx, "Encrypting installer file and constructing file metadata")
-			contentFile, encryptionInfo, err := encryptMobileAppAndConstructFileContentMetadata(ctx, installerSourcePath)
-			if err != nil {
-				resp.Diagnostics.AddError(
-					"Error constructing and encrypting Intune Mobile App content file",
-					err.Error(),
-				)
-				return
-			}
-
-			// Step 4d: Create the content file resource in Graph API
-			tflog.Debug(ctx, "Creating content file resource in Graph API")
-			createdFile, err := contentBuilder.
-				ByMobileAppContentId(*contentVersion.GetId()).
-				Files().
-				Post(ctx, contentFile, nil)
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-			tflog.Debug(ctx, fmt.Sprintf("Content file resource created with ID: %s", *createdFile.GetId()))
-
-			// Step 5d: Wait for Graph API to generate a valid Azure Storage URI
-			tflog.Debug(ctx, "Waiting for Graph API to generate a valid Azure Storage URI")
-			err = retry.RetryContext(ctx, time.Until(deadline), func() *retry.RetryError {
-				file, err := contentBuilder.
-					ByMobileAppContentId(*contentVersion.GetId()).
-					Files().
-					ByMobileAppContentFileId(*createdFile.GetId()).
-					Get(ctx, nil)
-				if err != nil {
-					tflog.Debug(ctx, fmt.Sprintf("Failed to get file status: %v", err))
-					return retry.RetryableError(fmt.Errorf("failed to get file status: %v", err))
-				}
-
-				if file.GetUploadState() == nil {
-					tflog.Debug(ctx, "Upload state is nil; retrying until a state is returned")
-					return retry.RetryableError(fmt.Errorf("upload state is nil"))
-				}
-
-				stateVal := *file.GetUploadState()
-				tflog.Debug(ctx, fmt.Sprintf("Current upload state: %s", stateVal.String()))
-
-				if stateVal == graphmodels.AZURESTORAGEURIREQUESTSUCCESS_MOBILEAPPCONTENTFILEUPLOADSTATE {
-					tflog.Debug(ctx, "Azure Storage URI request successful")
-					return nil
-				}
-
-				if stateVal == graphmodels.AZURESTORAGEURIREQUESTFAILED_MOBILEAPPCONTENTFILEUPLOADSTATE {
-					tflog.Debug(ctx, "Azure Storage URI request failed")
-					return retry.NonRetryableError(fmt.Errorf("azure storage URI request failed"))
-				}
-				tflog.Debug(ctx, fmt.Sprintf("Waiting for Azure Storage URI, current state: %s", stateVal.String()))
-				return retry.RetryableError(fmt.Errorf("waiting for Azure Storage URI, current state: %s", stateVal.String()))
-			})
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-
-			// Step 5e: Retrieve the Azure Storage URI and upload the encrypted file
-			tflog.Debug(ctx, "Retrieving file status for Azure Storage URI")
-			fileStatus, err := contentBuilder.
-				ByMobileAppContentId(*contentVersion.GetId()).
-				Files().
-				ByMobileAppContentFileId(*createdFile.GetId()).
-				Get(ctx, nil)
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-			if fileStatus.GetAzureStorageUri() == nil {
-				tflog.Debug(ctx, "Azure Storage URI is nil in the retrieved file status")
-				errors.HandleGraphError(ctx, fmt.Errorf("azure Storage URI is nil"), resp, "Update", r.WritePermissions)
-				return
-			}
-			tflog.Debug(ctx, fmt.Sprintf("Retrieved Azure Storage URI: %s", *fileStatus.GetAzureStorageUri()))
-
-			// IMPORTANT: Upload the encrypted file (.bin) not the original source file.
-			encryptedFilePath := installerSourcePath + ".bin"
-			tflog.Debug(ctx, fmt.Sprintf("Uploading encrypted file: %s", encryptedFilePath))
-			err = uploadToAzureStorage(ctx, *fileStatus.GetAzureStorageUri(), encryptedFilePath)
-			if err != nil {
-				tflog.Debug(ctx, fmt.Sprintf("Failed to upload to Azure Storage: %v", err))
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-
-			// Step 5f: Commit the file with encryption metadata
-			tflog.Debug(ctx, "Committing file with encryption metadata")
-			err = retry.RetryContext(ctx, time.Until(deadline), func() *retry.RetryError {
-				commitBody, err := CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
-				if err != nil {
-					return retry.NonRetryableError(fmt.Errorf("failed to construct commit request: %v", err))
-				}
-				err = contentBuilder.
-					ByMobileAppContentId(*contentVersion.GetId()).
-					Files().
-					ByMobileAppContentFileId(*createdFile.GetId()).
-					Commit().
-					Post(ctx, commitBody, nil)
-				if err != nil {
-					tflog.Debug(ctx, fmt.Sprintf("Failed to commit file: %v", err))
-					return retry.RetryableError(fmt.Errorf("failed to commit file: %v", err))
-				}
-				tflog.Debug(ctx, "File commit request successful")
-				return nil
-			})
-			if err != nil {
-				resp.Diagnostics.AddError("Error committing file", err.Error())
-				return
-			}
-
-			// Step 5g: Wait for commit to complete
-			tflog.Debug(ctx, "Waiting for file commit to complete")
-			maxRetries := 10
-			for i := 0; i < maxRetries; i++ {
-				file, err := contentBuilder.
-					ByMobileAppContentId(*contentVersion.GetId()).
-					Files().
-					ByMobileAppContentFileId(*createdFile.GetId()).
-					Get(ctx, nil)
-				if err != nil {
-					errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-					return
-				}
-
-				stateVal := *file.GetUploadState()
-				tflog.Debug(ctx, fmt.Sprintf("Commit status check %d: state=%s", i+1, stateVal.String()))
-				if stateVal == graphmodels.COMMITFILESUCCESS_MOBILEAPPCONTENTFILEUPLOADSTATE {
-					tflog.Debug(ctx, "File commit completed successfully")
-					break
-				}
-				if stateVal == graphmodels.COMMITFILEFAILED_MOBILEAPPCONTENTFILEUPLOADSTATE {
-					tflog.Debug(ctx, "File commit failed; retrying commit request")
-					commitBody, err := CommitUploadedMobileAppWithEncryptionMetadata(encryptionInfo)
-					if err != nil {
-						tflog.Debug(ctx, fmt.Sprintf("Error constructing commit request during retry: %v", err))
-						continue
-					}
-					err = contentBuilder.
-						ByMobileAppContentId(*contentVersion.GetId()).
-						Files().
-						ByMobileAppContentFileId(*createdFile.GetId()).
-						Commit().
-						Post(ctx, commitBody, nil)
-					if err != nil {
-						tflog.Debug(ctx, fmt.Sprintf("Error during commit retry: %v", err))
-						continue
-					}
-				}
-				if i == maxRetries-1 {
-					resp.Diagnostics.AddError(
-						"Error waiting for file commit",
-						fmt.Sprintf("File commit did not complete after %d attempts. Last state: %s", maxRetries, stateVal.String()),
-					)
-					return
-				}
-				time.Sleep(10 * time.Second)
-			}
-
-			// Step 5h: Update the App with the new Committed Content Version
-			updatePayload := graphmodels.NewMacOSPkgApp()
-			updatePayload.SetCommittedContentVersion(contentVersion.GetId())
-
-			_, err = r.client.
-				DeviceAppManagement().
-				MobileApps().
-				ByMobileAppId(object.ID.ValueString()).
-				Patch(ctx, updatePayload, nil)
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-			tflog.Debug(ctx, fmt.Sprintf("App updated with new committed content version: %s", *contentVersion.GetId()))
-		} else if existingContentVersion != "" {
-			tflog.Debug(ctx, fmt.Sprintf(
-				"File content unchanged, preserving existing content version: %s",
-				existingContentVersion,
-			))
-
-			// Ensure committedContentVersion still points to correct version
-			updatePayload := graphmodels.NewMacOSPkgApp()
-			updatePayload.SetCommittedContentVersion(&existingContentVersion)
-
-			_, err = r.client.
-				DeviceAppManagement().
-				MobileApps().
-				ByMobileAppId(object.ID.ValueString()).
-				Patch(ctx, updatePayload, nil)
-
-			if err != nil {
-				errors.HandleGraphError(ctx, err, resp, "Update", r.WritePermissions)
-				return
-			}
-
-			tflog.Debug(ctx, "Ensured existing content version is still committed")
-		}
-		// --- End File Processing Logic ---
-
-	}
+	// No update logic is defined here intentionally for the reupload of mobile apps as any changes to either
+	// installer_file_path_source orinstaller_url_source triggers a destory and redeploy. Implementation attempts were
+	// becoming grossly complex and overly difficult for not much benefit.
+	// Rationale
+	// Intune's default behaviour is to append new mobileAppContentFiles to a new mobileContainedApp versions. Incrementing by 1
+	// and references only the latest version.
+	// The Api docs suggest you can delete existing mobileContainedApp versions but you cannot if they are commited. Meaning
+	// there are disparities between the api docs and what was possible with the go sdk.
+	// https://learn.microsoft.com/en-us/graph/api/intune-apps-mobileappcontent-delete?view=graph-rest-1.0 <- doesnt work ?
+	//
+	// This causes terraform stating issues as we are not interesting in tracking previous versions, we only want the latest
+	// mobileContainedApp version and it's mobileAppContentFiles within.
+	// with no delete option this would cause stating issues.
+	//
+	// Since in real world scenarios' updating the mobile app content files would mean a new app deployment in real terms,
+	// as the only time you'd do this is if the app has installation bugs or unintended Ux. I believe force replacing
+	// the mobile app acheices the same outcome from a sys admin perspective.
+	//
+	// This approach ensures that the logic used to auto generate the detection logic for included_apps. Is always run and keeps
+	// the code more concise.
+	//
 
 	// Read updated resource state
 	err = retry.RetryContext(ctx, retryTimeout, func() *retry.RetryError {
